requirements.txt
appdirs==1.4.3
binaryornot==0.4.3
chardet==3.0.2
docopt==0.6.2
packaging==16.8
pyaml==16.12.2
pyparsing==2.2.0
PyYAML==3.12
six==1.10.0
